requirements.txt
django
django-braces
django-simple-history
docutils
python-dateutil
pytz
git+https://github.com/erikvw/django-crypto-fields@develop#egg=django_crypto_fields
git+https://github.com/erikvw/django-revision@develop#egg=django_revision
git+https://github.com/botswana-harvard/edc-constants@develop#egg=edc_constants
git+https://github.com/botswana-harvard/edc-device@develop#egg=edc-device
git+https://github.com/botswana-harvard/edc-protocol@develop#egg=edc_protocol
git+https://github.com/botswana-harvard/edc-base@develop#egg=edc_base
git+https://github.com/botswana-harvard/bais-subject-form-validators@develop#egg=bais_subject_form_validators